tests, cleanup, extra tests, split responsibility
